feat: ini tx components
