+++ packages/templates/src/transaction/transaction-info.tsx
